Fixed completely non-functional Cog Tournament backup. Fixed the VP never calling in new levels of backup. Backup now flies away instead of exploding when new backup is called in. The VP now says stuff when he calls in backup. The VP doesn't talk every time he throws an attack now. The VP no longer walks around (this caused issues with backup). Fixed Cog behaviors not being started in order of priorities. It's now possible to change the priorities of Cog behaviors on the fly. Fixed Cog behaviors not getting cleaned up when everyone leaves a battle (this caused crashes on the AI). Tournament backup no longer uses SuitHealBossBehavior, healing the boss is now part of SuitFollowBossBehavior to fix issues. Fixed backup not actually following the path they generated. Fixed backup generating a new path every time a checkpoint is reached, instead of just going to the next checkpoint in the path. There's now a delay of 8 seconds between announcing backup and backup actually spawning (6 seconds caused issues). Improved cog flying in and away animations. Improved cog tournament boss flying animation. Fixed slow ParticleAttack animations for Cogs. Improved the hop-on-trolley animation. Improved the camera angles for entering and exiting tunnels. Changed all uses of Whisper to WhisperPopup. Fixed an incorrect NPC in a building. Added JJ back to JJ's Diner. Made the TTC sky resemble the beta sky. Downscaled the TTC ground texture by 100% (this was so it would look more like the original). Fixed incorrect path to a 'sit' animation for Toon legs. Changed the Toon movement speed to match old Toontown. You can now view the DOIDs of Cogs. Added the original .dna file of new_ttc_sz back so it can be edited.

+++ game/lib/coginvasion/cog/DistributedSuitAI.py
@@ -277,8 +277,9 @@ class DistributedSuitAI(DistributedAvatarAI, DistributedSmoothNodeAI):
             self.track = Sequence(Wait(6.0), Func(self.closeSuit))
             self.track.start()
 
-    def closeSuit(self):
-        self.itemDropper.drop()
+    def closeSuit(self, dropItem = True):
+        if dropItem:
+            self.itemDropper.drop()
         if self.getManager():
             self.getManager().deadSuit(self.doId)
         self.disable()
@@ -292,25 +293,27 @@ class DistributedSuitAI(DistributedAvatarAI, DistributedSmoothNodeAI):
         self.brain.addBehavior(SuitAttackBehavior(self), priority = 3)
         if self.suitPlan.getName() != SuitGlobals.VicePresident:
             self.brain.addBehavior(SuitPanicBehavior(self), priority = 2)
+            self.brain.addBehavior(SuitRandomStrollBehavior(self), priority = 1)
         else:
             self.brain.addBehavior(SuitFlyToRandomSpotBehavior(self), priority = 2)
             self.brain.addBehavior(SuitCallInBackupBehavior(self), priority = 4)
-        self.brain.addBehavior(SuitRandomStrollBehavior(self), priority = 1)
         place = CIGlobals.SuitSpawnPoints[self.hood]
         landspot = random.choice(place.keys())
         path = place[landspot]
         index = place.keys().index(landspot)
         self.currentPath = landspot
-        self.setH(random.uniform(0.0, 360.0))
+        yaw = random.uniform(0.0, 360.0)
+        self.setH(yaw)
+        self.d_setH(yaw)
         self.clearTrack()
         self.track = Sequence()
         if spawnMode == SpawnMode.FLYDOWN:
-            flyTrack = self.posInterval(3,
-                path, startPos = path + (0, 0, 50)
+            flyTrack = self.posInterval(3.5,
+                path, startPos = path + (0, 0, 6.5 * 4.8)
             )
             flyTrack.start()
             self.b_setSuitState(2, index, index)
-            self.track.append(Wait(5.4))
+            self.track.append(Wait(6.5))
         self.track.append(Func(self.b_setAnimState, 'neutral'))
         self.track.append(Wait(1.0))
         self.track.append(Func(self.brain.startThinking))
@@ -348,6 +351,7 @@ class DistributedSuitAI(DistributedAvatarAI, DistributedSmoothNodeAI):
         taskMgr.remove(self.uniqueName('monitorHealth'))
         if self.brain:
             self.brain.stopThinking()
+            self.brain.unloadBehaviors()
             self.brain = None
         self.itemDropper.cleanup()
         self.itemDropper = None
